restructured rate_control
